+++ eabm/eabm_package/python/tutorial_07_latin_hypercube.py
@@ -1,0 +1,13 @@
+parameter_schema = {
+    'num_simulations': 4,
+    'width': {
+        'distribution': 'norm',
+        'loc': 1,
+        'scale': .1
+    },
+    'height': {
+        'distribution': 'norm',
+        'loc': 1,
+        'scale': .1
+    }
+}